+++ docutils/test/test_parsers/test_rst/test_directives/test_images.py
@@ -369,7 +369,7 @@ totest['images'] = [
 """,
 """\
 <document source="test data">
-    <reference name="indirect" refname="indirect">
+    <reference refname="indirect">
         <image uri="picture.png">
 """],
 ["""\
@@ -385,7 +385,7 @@ totest['images'] = [
 <document source="test data">
     <reference refuri="a/multi/line/uri">
         <image uri="picture.png">
-    <reference name="a multi line internal reference" refname="a multi line internal reference">
+    <reference refname="a multi line internal reference">
         <image uri="picture.png">
 """],
 ["""\
@@ -475,7 +475,7 @@ totest['images'] = [
 """,
 """\
 <document source="test data">
-    <reference name="Uppercase" refname="uppercase">
+    <reference refname="uppercase">
         <image uri="test.png">
     <target ids="uppercase" names="uppercase" refuri="https://docutils.sourceforge.io/">
 """],