+++ docutils/test/test_transforms/test_contents.py
@@ -99,7 +99,7 @@ Paragraph 4.
             Paragraph 1.
         <section ids="title-2" names="title\\ 2">
             <title>
-                <reference name="Title" refname="title">
+                <reference refname="title">
                     Title
                  2
             <paragraph>